Basic processing for Fast2DC.  Still some issues with "live-time" of a particles, or dead time for the probe, depending on how you want to look at it.  This will provide the processing engine for nimbus.
git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk@3865 640d5228-2204-0410-8d71-8f46fa6850b4

+++ xpms2d/class/Hex.h
@@ -24,7 +24,7 @@ public:
 
 	Hex(const Widget parent);
 
-  void	Update(int nBuffs, P2d_rec sets[]);
+  void	Update(size_t nBuffs, P2d_rec sets[]);
 
 private:
